+++ src/components/Polystat.tsx
@@ -444,6 +444,7 @@ export const Polystat: React.FC<PolystatOptions> = (options) => {
     return (
       <text
         className="toplabel"
+        data-testid={`polystat-label-${index}`}
         x={coords.x + alignments.labelTextAlignmentX}
         y={coords.y + verticalAlignment}
         textAnchor="middle"
@@ -509,6 +510,7 @@ export const Polystat: React.FC<PolystatOptions> = (options) => {
       <text
         ref={animationRefs[index]}
         className={`valueLabel${index}`}
+        data-testid={`polystat-value-${index}`}
         x={coords.x + alignments.labelValueAlignmentX}
         y={coords.y + verticalAlignment}
         textAnchor="middle"
@@ -553,6 +555,7 @@ export const Polystat: React.FC<PolystatOptions> = (options) => {
       <text
         ref={animationTimestampRefs[index]}
         className={`timestampLabel${index}`}
+        data-testid={`polystat-timestamp-${index}`}
         x={coords.x + alignments.labelValueAlignmentX}
         y={coords.y + verticalAlignment}
         textAnchor="middle"